chore: move dashboard interface in dashboard/
