+++ mergify_engine/actions/merge.py
@@ -27,8 +27,8 @@ from mergify_engine import context
 from mergify_engine import queue
 from mergify_engine import rules
 from mergify_engine import signals
-from mergify_engine import subscription
 from mergify_engine.actions import merge_base
+from mergify_engine.dashboard import subscription
 from mergify_engine.queue import naive
 from mergify_engine.rules import types
