+++ mergify_engine/actions/merge_base.py
@@ -31,11 +31,11 @@ from mergify_engine import github_types
 from mergify_engine import json as mergify_json
 from mergify_engine import queue
 from mergify_engine import rules
-from mergify_engine import subscription
-from mergify_engine import user_tokens
 from mergify_engine import utils
 from mergify_engine.actions import utils as action_utils
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
+from mergify_engine.dashboard import user_tokens
 from mergify_engine.rules import conditions
