+++ mergify_engine/actions/post_check.py
@@ -23,7 +23,7 @@ from mergify_engine import check_api
 from mergify_engine import context
 from mergify_engine import rules
 from mergify_engine import signals
-from mergify_engine import subscription
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import types
