+++ mergify_engine/actions/queue.py
@@ -26,9 +26,9 @@ from mergify_engine import delayed_refresh
 from mergify_engine import github_types
 from mergify_engine import queue
 from mergify_engine import signals
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.actions import merge_base
+from mergify_engine.dashboard import subscription
 from mergify_engine.queue import merge_train
 from mergify_engine.rules import types
