+++ mergify_engine/actions/rebase.py
@@ -23,8 +23,8 @@ from mergify_engine import check_api
 from mergify_engine import context
 from mergify_engine import rules
 from mergify_engine import signals
-from mergify_engine import subscription
 from mergify_engine.actions import utils as action_utils
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import conditions
 from mergify_engine.rules import types
