+++ mergify_engine/actions/request_reviews.py
@@ -21,9 +21,9 @@ from mergify_engine import check_api
 from mergify_engine import context
 from mergify_engine import rules
 from mergify_engine import signals
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import types
