+++ mergify_engine/actions/utils.py
@@ -22,8 +22,8 @@ import voluptuous
 from mergify_engine import check_api
 from mergify_engine import context
 from mergify_engine import github_types
-from mergify_engine import subscription
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import types
