+++ mergify_engine/queue/merge_train.py
@@ -33,10 +33,10 @@ from mergify_engine import github_types
 from mergify_engine import json
 from mergify_engine import queue
 from mergify_engine import rules
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.actions import merge_base
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 
 
 CHECK_ASSERTS = {
