+++ mergify_engine/tests/functional/actions/test_post_check.py
@@ -19,8 +19,8 @@ import yaml
 
 from mergify_engine import config
 from mergify_engine import context
-from mergify_engine import subscription
 from mergify_engine import utils
+from mergify_engine.dashboard import subscription
 from mergify_engine.tests.functional import base
