+++ mergify_engine/tests/functional/test_debug.py
@@ -21,7 +21,7 @@ import yaml
 from mergify_engine import config
 from mergify_engine import context
 from mergify_engine import debug
-from mergify_engine import subscription
+from mergify_engine.dashboard import subscription
 from mergify_engine.tests.functional import base
