+++ mergify_engine/tests/functional/test_engine.py
@@ -1178,7 +1178,7 @@ class TestEngineV2Scenario(base.FunctionalTestBase):
 
     async def test_marketplace_event(self):
         with mock.patch(
-            "mergify_engine.subscription.Subscription.get_subscription"
+            "mergify_engine.dashboard.subscription.Subscription.get_subscription"
         ) as get_sub:
             get_sub.return_value = self.subscription
             r = await self.app.post(
