+++ mergify_engine/tests/functional/test_subscription.py
@@ -17,7 +17,7 @@
 from mergify_engine import config
 from mergify_engine import constants
 from mergify_engine import context
-from mergify_engine import subscription
+from mergify_engine.dashboard import subscription
 from mergify_engine.tests.functional import base
