+++ mergify_engine/tests/unit/actions/test_merge.py
@@ -21,9 +21,9 @@ import voluptuous
 from mergify_engine import constants
 from mergify_engine import context
 from mergify_engine import github_types
-from mergify_engine import subscription
 from mergify_engine.actions import merge
 from mergify_engine.actions import merge_base
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import conditions
