+++ mergify_engine/tests/unit/actions/test_request_reviews.py
@@ -19,9 +19,9 @@ import voluptuous
 from mergify_engine import check_api
 from mergify_engine import context
 from mergify_engine import github_types
-from mergify_engine import subscription
 from mergify_engine.actions import request_reviews
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 
 
 @pytest.mark.parametrize(
