+++ mergify_engine/tests/unit/rules/test_rules.py
@@ -27,10 +27,10 @@ from mergify_engine import context
 from mergify_engine import date
 from mergify_engine import github_types
 from mergify_engine import rules
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.actions import merge_base
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 from mergify_engine.rules import InvalidRules
 from mergify_engine.rules import conditions
 from mergify_engine.rules import get_mergify_config
