+++ mergify_engine/tests/unit/test_context.py
@@ -21,10 +21,10 @@ import pytest
 from mergify_engine import config
 from mergify_engine import context
 from mergify_engine import github_types
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.clients import github
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 
 
 @pytest.mark.asyncio
