+++ mergify_engine/tests/unit/test_duplicate.py
@@ -21,8 +21,8 @@ import pytest
 from mergify_engine import context
 from mergify_engine import duplicate_pull
 from mergify_engine import github_types
-from mergify_engine import subscription
 from mergify_engine import utils
+from mergify_engine.dashboard import subscription
 
 
 async def fake_get_github_pulls_from_sha(url, api_version=None):
