+++ mergify_engine/tests/unit/test_signals.py
@@ -19,7 +19,7 @@ import pytest
 from mergify_engine import context
 from mergify_engine import github_types
 from mergify_engine import signals
-from mergify_engine import subscription
+from mergify_engine.dashboard import subscription
 
 
 @pytest.mark.asyncio
