+++ mergify_engine/tests/unit/test_train.py
@@ -28,8 +28,8 @@ from mergify_engine import context
 from mergify_engine import github_types
 from mergify_engine import queue
 from mergify_engine import rules
-from mergify_engine import subscription
 from mergify_engine.actions import merge_base
+from mergify_engine.dashboard import subscription
 from mergify_engine.queue import merge_train
