+++ mergify_engine/web/root.py
@@ -32,11 +32,11 @@ from mergify_engine import count_seats
 from mergify_engine import github_events
 from mergify_engine import github_types
 from mergify_engine import json
-from mergify_engine import subscription
-from mergify_engine import user_tokens
 from mergify_engine import utils
 from mergify_engine.clients import github
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
+from mergify_engine.dashboard import user_tokens
 from mergify_engine.queue import merge_train
 from mergify_engine.web import auth
 from mergify_engine.web import badges
