+++ mergify_engine/web/simulator.py
@@ -27,10 +27,10 @@ from mergify_engine import context
 from mergify_engine import exceptions
 from mergify_engine import github_types
 from mergify_engine import rules
-from mergify_engine import subscription
 from mergify_engine import utils
 from mergify_engine.clients import github
 from mergify_engine.clients import http
+from mergify_engine.dashboard import subscription
 from mergify_engine.engine import actions_runner
 from mergify_engine.web import auth
 from mergify_engine.web import redis
